--- src/components/board/types.tsx
@@ -1,4 +1,0 @@
-export type boardType = { [key: string]: { color: string, type: string, position: string } };
-
-
-export type boardTypeWithoutMoving = { [key: string]: { color: string, type: string, canMove: Function } };
