--- src/components/figures/Figures.module.scss
@@ -1,390 +1,0 @@
-.figure {
-    position: absolute;
-
-    width: 100px;
-    height: 100px;
-
-    background-size: 100px;
-
-    z-index: 1;
-}
-
-.king-black {
-    background-image: url('./../../images/Chess_tile_kd.svg');
-}
-
-.king-white {
-    background-image: url('./../../images/Chess_tile_kl-whitebg.svg');
-}
-
-.queen-black {
-    background-image: url('./../../images/Chess_tile_qd.svg');
-}
-
-.queen-white {
-    background-image: url('./../../images/Chess_tile_ql-whitebg.svg');
-}
-
-.rook-black {
-    background-image: url('./../../images/Chess_tile_rd.svg');
-}
-
-.rook-white {
-    background-image: url('./../../images/whitebg.svg');
-}
-
-.knight-black {
-    background-image: url('./../../images/Chess_tile_nd.svg');
-}
-
-.knight-white {
-    background-image: url('./../../images/Chess_tile_nl.svg');
-}
-
-.bishop-black {
-    /* background-image: url('./images/Chess_tile_bd.svg'); */
-
-    background-image: url("data:image/svg+xml,%3Csvg xmlns:svg='http://www.w3.org/2000/svg' xmlns='http://www.w3.org/2000/svg' version='1.0' width='64' height='64' viewBox='0 0 504 504' id='svg1923'%3E%3Cdefs id='defs1926' /%3E%3Cpath d='M 277.97634 79.998123 A 26.390158 26.390158 0 1 1 225.19602,79.998123 A 26.390158 26.390158 0 1 1 277.97634 79.998123 z' transform='matrix(0.966666,0,0,0.966666,9.265884,2.758784e-2)' style='fill:%23000000;fill-opacity:1;stroke:none;stroke-width:0.2;stroke-miterlimit:4;stroke-dasharray:none' id='path1932' /%3E%3Cpath d='M 253.10322,101.98992 C 253.10322,101.98992 338.21149,160.48811 349.86714,180.28073 C 376.81239,226.03681 378.56624,271.31555 327.87534,300.79578 C 326.93756,301.34116 343.51707,358.7205 341.95009,359.7338 C 302.11708,385.49203 266.2983,383.48494 266.2983,383.48494 C 266.2983,383.48494 279.49338,420.871 374.49795,410.75477 C 424.15679,405.51553 456.30744,445.06198 456.30744,445.06198 L 429.03761,474.09115 C 429.03761,474.09115 395.41423,445.18704 349.86714,447.701 C 282.1324,451.43961 253.10322,425.7092 253.10322,425.7092 C 253.10322,425.7092 217.95756,452.07518 165.13603,447.701 C 108.63961,443.0225 79.807839,474.97083 79.807839,474.97083 L 49.019319,445.94165 C 49.019319,445.94165 92.598879,402.25358 156.33931,409.8751 C 223.40221,417.89389 239.90815,386.12396 239.90815,386.12396 C 239.90815,386.12396 194.85528,379.3026 164.25636,359.7338 C 162.90757,358.87121 178.33111,300.79578 178.33111,300.79578 C 178.33111,300.79578 98.281789,256.13589 160.73767,174.12302 C 177.03775,152.71885 253.10322,101.98992 253.10322,101.98992 z ' style='fill:%23000000;fill-opacity:1;fill-rule:evenodd;stroke:none;stroke-width:1.79999995;stroke-linecap:butt;stroke-linejoin:miter;stroke-miterlimit:4;stroke-dasharray:none;stroke-opacity:1' id='path1934' /%3E%3Crect width='15.174341' height='65.315643' x='244.98863' y='173.90311' style='fill:%23ffffff;fill-opacity:1;stroke:none;stroke-width:1.79999995;stroke-miterlimit:4;stroke-dasharray:none;stroke-opacity:1' id='rect5478' /%3E%3Crect width='63.996132' height='14.954423' x='219.91798' y='198.97375' style='fill:%23ffffff;fill-opacity:1;stroke:none;stroke-width:1.79999995;stroke-miterlimit:4;stroke-dasharray:none;stroke-opacity:1' id='rect5480' /%3E%3Cpath d='M 187.80996,295.07791 C 223.15724,272.03508 286.77531,273.453 316.24206,294.63808 C 316.40301,294.7538 310.18969,307.18019 309.86444,306.95349 C 285.67395,290.09315 224.11436,288.59218 195.28717,306.95349 C 195.0051,307.13315 187.57374,295.2319 187.80996,295.07791 z ' style='fill:%23ffffff;fill-opacity:1;fill-rule:evenodd;stroke:none;stroke-width:1px;stroke-linecap:butt;stroke-linejoin:miter;stroke-opacity:1' id='path6367' /%3E%3Cpath d='M 171.38828,355.09632 C 216.63197,316.65937 293.88485,319.8365 331.9284,354.65649 C 332.07463,354.79033 323.88082,366.34025 323.57152,366.09223 C 292.12374,340.87501 217.31903,332.43566 180.84475,366.31215 C 180.63598,366.50605 171.17338,355.27889 171.38828,355.09632 z ' style='fill:%23ffffff;fill-opacity:1;fill-rule:evenodd;stroke:none;stroke-width:1px;stroke-linecap:butt;stroke-linejoin:miter;stroke-opacity:1' id='path8139' /%3E%3C/svg%3E%0A");
-}
-
-.bishop-white {
-    background-image: url('./../../images/Chess_tile_bl.svg');
-}
-
-.pawn-black {
-    background-image: url('./../../images/Chess_tile_pd.svg');
-}
-
-.pawn-white {
-    background-image: url('./../../images/Chess_tile_pl.svg');
-}
-
-/* Positions */
-
-.positionA8 {
-    left: 0px;
-    top: 0px;
-}
-
-.positionA7 {
-    left: 0px;
-    top: 100px;
-}
-
-.positionA6 {
-    left: 0px;
-    top: 200px;
-}
-
-.positionA5 {
-    left: 0px;
-    top: 300px;
-}
-
-.positionA4 {
-    left: 0px;
-    top: 400px;
-}
-
-.positionA3 {
-    left: 0px;
-    top: 500px;
-}
-
-.positionA2 {
-    left: 0px;
-    top: 600px;
-}
-
-.positionA1 {
-    left: 0px;
-    top: 700px;
-}
-
-.positionB8 {
-    left: 100px;
-    top: 0px;
-}
-
-.positionB7 {
-    left: 100px;
-    top: 100px;
-}
-
-.positionB6 {
-    left: 100px;
-    top: 200px;
-}
-
-.positionB5 {
-    left: 100px;
-    top: 300px;
-}
-
-.positionB4 {
-    left: 100px;
-    top: 400px;
-}
-
-.positionB3 {
-    left: 100px;
-    top: 500px;
-}
-
-.positionB2 {
-    left: 100px;
-    top: 600px;
-}
-
-.positionB1 {
-    left: 100px;
-    top: 700px;
-}
-
-.positionC8 {
-    left: 200px;
-    top: 0px;
-}
-
-.positionC7 {
-    left: 200px;
-    top: 100px;
-}
-
-.positionC6 {
-    left: 200px;
-    top: 200px;
-}
-
-.positionC5 {
-    left: 200px;
-    top: 300px;
-}
-
-.positionC4 {
-    left: 200px;
-    top: 400px;
-}
-
-.positionC3 {
-    left: 200px;
-    top: 500px;
-}
-
-.positionC2 {
-    left: 200px;
-    top: 600px;
-}
-
-.positionC1 {
-    left: 200px;
-    top: 700px;
-}
-
-.positionD8 {
-    left: 300px;
-    top: 0px;
-}
-
-.positionD7 {
-    left: 300px;
-    top: 100px;
-}
-
-.positionD6 {
-    left: 300px;
-    top: 200px;
-}
-
-.positionD5 {
-    left: 300px;
-    top: 300px;
-}
-
-.positionD4 {
-    left: 300px;
-    top: 400px;
-}
-
-.positionD3 {
-    left: 300px;
-    top: 500px;
-}
-
-.positionD2 {
-    left: 300px;
-    top: 600px;
-}
-
-.positionD1 {
-    left: 300px;
-    top: 700px;
-}
-
-.positionE8 {
-    left: 400px;
-    top: 0px;
-}
-
-.positionE7 {
-    left: 400px;
-    top: 100px;
-}
-
-.positionE6 {
-    left: 400px;
-    top: 200px;
-}
-
-.positionE5 {
-    left: 400px;
-    top: 300px;
-}
-
-.positionE4 {
-    left: 400px;
-    top: 400px;
-}
-
-.positionE3 {
-    left: 400px;
-    top: 500px;
-}
-
-.positionE2 {
-    left: 400px;
-    top: 600px;
-}
-
-.positionE1 {
-    left: 400px;
-    top: 700px;
-}
-
-.positionF8 {
-    left: 500px;
-    top: 0px;
-}
-
-.positionF7 {
-    left: 500px;
-    top: 100px;
-}
-
-.positionF6 {
-    left: 500px;
-    top: 200px;
-}
-
-.positionF5 {
-    left: 500px;
-    top: 300px;
-}
-
-.positionF4 {
-    left: 500px;
-    top: 400px;
-}
-
-.positionF3 {
-    left: 500px;
-    top: 500px;
-}
-
-.positionF2 {
-    left: 500px;
-    top: 600px;
-}
-
-.positionF1 {
-    left: 500px;
-    top: 700px;
-}
-
-.positionG8 {
-    left: 600px;
-    top: 0px;
-}
-
-.positionG7 {
-    left: 600px;
-    top: 100px;
-}
-
-.positionG6 {
-    left: 600px;
-    top: 200px;
-}
-
-.positionG5 {
-    left: 600px;
-    top: 300px;
-}
-
-.positionG4 {
-    left: 600px;
-    top: 400px;
-}
-
-.positionG3 {
-    left: 600px;
-    top: 500px;
-}
-
-.positionG2 {
-    left: 600px;
-    top: 600px;
-}
-
-.positionG1 {
-    left: 600px;
-    top: 700px;
-}
-
-.positionH8 {
-    left: 700px;
-    top: 0px;
-}
-
-.positionH7 {
-    left: 700px;
-    top: 100px;
-}
-
-.positionH6 {
-    left: 700px;
-    top: 200px;
-}
-
-.positionH5 {
-    left: 700px;
-    top: 300px;
-}
-
-.positionH4 {
-    left: 700px;
-    top: 400px;
-}
-
-.positionH3 {
-    left: 700px;
-    top: 500px;
-}
-
-.positionH2 {
-    left: 700px;
-    top: 600px;
-}
-
-.positionH1 {
-    left: 700px;
-    top: 700px;
-}
-
-.clicked:not(.beaten) {
-    background-color: yellow;
-}
-
-.beaten {
-    display: none;
-}
