+++ src/main.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.css";
-import { NavigationMenu } from "./components/navigationMenu/index.tsx";
-import Board from "./components/board/Board.tsx";
+import { NavigationMenu } from "./screens/navigationMenu/index.tsx";
+import Board from "./screens/board/Board.tsx";
 import { Provider } from "react-redux";
 import store from "./redux/store.tsx";
