+++ src/screens/board/config.json
