+++ src/screens/board/generateFigures.tsx
