+++ src/screens/gameOverModal/GameOverModal.module.scss
