+++ src/screens/navigationMenu/NavigationMenu.module.scss
