+++ src/shared/canMove/canMoveQueen.tsx
@@ -1,6 +1,6 @@
-import { boardType } from "../../board/types";
-import canMoveBishop from "../bishop/canMoveBishop";
-import canMoveRook from "../rook/canMoveRook";
+import { boardType } from "../constants/_types";
+import canMoveBishop from "./canMoveBishop";
+import canMoveRook from "./canMoveRook";
 
 export default function canMoveQueen(color: string, from: string, board: boardType, recurse = true) {
     const fromRook = canMoveRook(color, from, board, recurse);
