+++ src/shared/constants/_types.tsx
@@ -1,11 +1,7 @@
 export type boardType = {
   [key: string]: {
-    moveTo: Function;
     color: string;
-    canMove: Function;
     type: string;
+    position: string;
   };
-};
-export type boardTypeWithoutMoving = {
-  [key: string]: { color: string; type: string; canMove: Function };
 };
